refactor : 리퀘스트 리팩토링
