+++ tomcat/src/main/java/org/apache/coyote/http11/Http11Processor.java
@@ -1,17 +1,12 @@
 package org.apache.coyote.http11;
 
-import java.io.BufferedReader;
 import java.io.File;
 import java.io.IOException;
-import java.io.InputStream;
-import java.io.InputStreamReader;
 import java.net.Socket;
 import java.net.URL;
 import java.nio.file.Files;
-import java.util.HashMap;
 import java.util.Map;
 import java.util.Optional;
-import java.util.StringTokenizer;
 
 import org.apache.coyote.Processor;
 import org.slf4j.Logger;
@@ -43,7 +38,7 @@ public class Http11Processor implements Runnable, Processor {
         try (final var inputStream = connection.getInputStream();
              final var outputStream = connection.getOutputStream()) {
 
-            HttpRequest request = createHttpRequestFrom(inputStream);
+            HttpRequest request = HttpRequestParser.from(inputStream).toHttpRequest();
 
             HttpResponse httpResponse = createHttpResponseFrom(request);
 
@@ -52,52 +47,6 @@ public class Http11Processor implements Runnable, Processor {
         } catch (IOException | UncheckedServletException e) {
             log.error(e.getMessage(), e);
         }
-    }
-
-    private HttpRequest createHttpRequestFrom(InputStream inputStream) throws IOException {
-        BufferedReader bufferedReader = new BufferedReader(new InputStreamReader(inputStream));
-
-        String requestLine = bufferedReader.readLine();
-        StringTokenizer requestTokenizer = new StringTokenizer(requestLine);
-        HttpMethod httpMethod = HttpMethod.of(requestTokenizer.nextToken());
-        String uri = requestTokenizer.nextToken();
-
-        Map<String, String> httpRequestHeaders = new HashMap<>();
-        String line = bufferedReader.readLine();
-        while (!"".equals(line)) {
-            StringTokenizer headerTokenizer =
-                new StringTokenizer(line.replaceAll(" ", "").replaceAll(":", " "));
-            httpRequestHeaders.put(headerTokenizer.nextToken(), headerTokenizer.nextToken());
-            line = bufferedReader.readLine();
-        }
-
-        Map<String, String> cookies = new HashMap<>();
-        String cookie = httpRequestHeaders.getOrDefault("Cookie", "")
-            .replaceAll(" ", "")
-            .replaceAll(";", " ");
-        StringTokenizer cookieTokenizer = new StringTokenizer(cookie);
-        while (cookieTokenizer.hasMoreTokens()) {
-            String attribute = cookieTokenizer.nextToken().replaceAll("=", " ");
-            StringTokenizer attributeTokenizer = new StringTokenizer(attribute);
-            cookies.put(attributeTokenizer.nextToken(), attributeTokenizer.nextToken());
-        }
-
-        Map<String, String> bodyParams = new HashMap<>();
-        int contentLength = Integer.parseInt(httpRequestHeaders.getOrDefault("Content-Length", "0"));
-        char[] buffer = new char[contentLength];
-        bufferedReader.read(buffer, 0, contentLength);
-        StringTokenizer bodyTokenizer = new StringTokenizer(new String(buffer).replaceAll("&", " "));
-        while (bodyTokenizer.hasMoreTokens()) {
-            String attribute = bodyTokenizer.nextToken().replaceAll("=", " ");
-            StringTokenizer attributeTokenizer = new StringTokenizer(attribute);
-            bodyParams.put(attributeTokenizer.nextToken(), attributeTokenizer.nextToken());
-        }
-
-        return HttpRequestBuilder.of(httpMethod, uri, HttpStatusCode.HTTP_STATUS_OK)
-            .setRequestHeaders(httpRequestHeaders)
-            .setHttpCookie(cookies)
-            .setBodyParams(bodyParams)
-            .toHttpRequest();
     }
 
     private HttpResponse createHttpResponseFrom(HttpRequest request) throws IOException {
@@ -113,13 +62,7 @@ public class Http11Processor implements Runnable, Processor {
             return loginUser(request);
         }
 
-        final URL resource = getClass().getClassLoader().getResource(request.getStaticPath());
-        String responseBody = new String(Files.readAllBytes(new File(resource.getFile()).toPath()));
-
-        return HttpResponse.of(request.getHttpStatusCode(), responseBody)
-            .setContentType(request.getContentType() + ";charset=utf-8")
-            .setLocation(request.getPath())
-            .setCookie(request.getCookie().toString());
+        return createGetResponseFrom(request);
     }
 
     private HttpResponse loginWithAuthorization(HttpRequest request) throws IOException {
@@ -132,24 +75,23 @@ public class Http11Processor implements Runnable, Processor {
             return redirectTo("/index", HttpStatusCode.HTTP_STATUS_FOUND)
                 .setCookie(httpCookie.toString());
         }
-
-        final URL resource = getClass().getClassLoader().getResource(request.getStaticPath());
-        String responseBody = new String(Files.readAllBytes(new File(resource.getFile()).toPath()));
-
-        return HttpResponse.of(request.getHttpStatusCode(), responseBody)
-            .setContentType(request.getContentType() + ";charset=utf-8")
-            .setLocation(request.getPath());
+        return createGetResponseFrom(request);
     }
 
     private HttpResponse loginUser(HttpRequest request) throws IOException {
         Optional<User> user = InMemoryUserRepository.findByAccount(request.getBodyParam("account"));
 
-        if (isAuthorized(request, user)) {
+        if (user.isEmpty()) {
+            return redirectTo("/401", HttpStatusCode.HTTP_STATUS_UNAUTHORIZED);
+        }
+
+        User authorizedUser = user.get();
+        if (authorizedUser.checkPassword(request.getBodyParam("password"))) {
             return redirectTo("/401", HttpStatusCode.HTTP_STATUS_UNAUTHORIZED);
         }
 
         Session session = request.getSession();
-        session.setAttribute("user", user.get());
+        session.setAttribute("user", authorizedUser);
         SessionManager.getSessionManager().add(session);
 
         Map<String, String> cookies = Map.of("JSESSIONID", session.getId());
@@ -157,11 +99,6 @@ public class Http11Processor implements Runnable, Processor {
 
         return redirectTo("/index", HttpStatusCode.HTTP_STATUS_FOUND)
             .setCookie(httpCookie.toString());
-    }
-
-    private boolean isAuthorized(HttpRequest request, Optional<User> user) {
-        return user.isPresent()
-            && user.get().checkPassword(request.getBodyParam("password"));
     }
 
     private HttpResponse registerUser(HttpRequest request) throws IOException {
@@ -172,11 +109,22 @@ public class Http11Processor implements Runnable, Processor {
     }
 
     private HttpResponse redirectTo(String location, HttpStatusCode httpStatusCode) throws IOException {
-        final URL resource = getClass().getClassLoader().getResource("static" + location + ".html");
+        return createGetResponseFrom("static" + location + ".html", httpStatusCode, "text/html", location);
+    }
+
+    private HttpResponse createGetResponseFrom(HttpRequest request) throws IOException {
+        return createGetResponseFrom(request.getResourcePath(), HttpStatusCode.HTTP_STATUS_OK, request.getContentType(),
+            request.getPath());
+    }
+
+    private HttpResponse createGetResponseFrom(String resourcePath, HttpStatusCode httpStatusCode, String ContentType,
+        String location) throws
+        IOException {
+        final URL resource = getClass().getClassLoader().getResource(resourcePath);
         String responseBody = new String(Files.readAllBytes(new File(resource.getFile()).toPath()));
 
         return HttpResponse.of(httpStatusCode, responseBody)
-            .setContentType("text/html;charset=utf-8")
+            .setContentType(ContentType + ";charset=utf-8")
             .setLocation(location);
     }
 }
